# Limit mempool by throwing away the cheapest txn and setting min relay fee to it