Tests forthcoming, but I felt bad I still hadnt pushed this.
See commitmsg for more details.