@JeremyRubin 2015-09-25
I think that my earlier comment is not fully needed, because mempool is a large multiple of block size, currently. Perhaps a more future proof implementation would allow setting:

- an optional hard memory cap
- a (potentially) dynamic size which is a large multiple of the current block size