@morcos 2015-09-25
@TheBlueMatt re: my comment on high fee txs.  I see now, you aren't doing the overall fee check in order to boot a package.  I just assumed the StageTrimToSize logic was the same.  So how do you think about free relay then?  Could you write up a quick intro describing the algorithm as it would help to know how you think about it.  Is the idea that all even though the tx causing the eviction hasn't covered the fees to pay for the evicted packages relay, by boosting the minRelayRate you're essentially forcing all future transactions to do so?

It's an interesting idea, one question is how big a sweet spot there is between having the half-life too long  and worrying about the "cram relayFee high all of a sudden" attack vs having it too low and perhaps having some vague concern about free relay.

Why does your increased relay fee only apply to low priority transactions?  I think it has to apply to all.