@morcos 2015-09-26
But in particular the increased relay fee does NOT apply to high priority txs?  That's what I don't understand.  It seems you could use the same stable of high priority inputs over and over to gain free relay.