@NanoAkron 2015-09-26
What't wrong with XT's method of discarding a random transaction so that you can't predictably  manipulate the mempool?