@TheBlueMatt 2015-09-26
@NanoAkron It makes it trivial to DoS the network, among many other issues.