@morcos 2015-09-30
I think the `ROLLING_FEE_HALFLIFE` should be 12 hours.  Here's my analysis:
The purpose of the `rollingMinimumFeeRate` is to strike the right balance between two things.  
- Future transactions should be obligated to pay for the cost of transactions that were evicted (and their own relay fee) otherwise a large package of transactions could be evicted by a small tx with a slightly higher fee rate.  This could happen repeatedly for a bandwidth attack.
- It must decay so an attacker can not pack the mempool full of high fee txs one time and peg the effective min relay rate very high for a long time for the cost of stuffing the mempool once.

From the point of view of the bandwidth attack:  
Assume the prevailing fee rate at the bottom of the mempool is X times the relay rate.  Then a full size 2.5MB package can be evicted from there by paying X+1 on a small 200 byte tx.  Effectively you have now paid the minimum relay fee on (200X + 200) bytes, but have relayed 2.5MB + 200 bytes, so you got free relay of 2.5MB - X * 200 bytes.  

As soon as the `rollingMinimumFeeRate` has dropped from X back down to X-1, you can repeat the attack.  At a half-life of 12 hours and assuming X = 20, then it'll take about 53 mins for that to happen.  So you can free relay 47 kB per min.  This seems sufficiently small compared to the bare minimum network relay capacity of 100 kB per min (1 block every 10 mins).  

Since the decay is exponential, you'll actually take a lot longer than 53 mins to repeat the attack if the prevailing fee rate multiple X is considerably less than 20.   However as the prevailing fee rate climbs the attack could be considered a bigger concern.  This should be addressed by having a default minimum relay rate that is higher.  It seams reasonable that over the long term the default minimum relay rate will not be much less than 1/20th of the prevailing fee rate at the bottom of mempools.


From the point of view of stuffing the mempool:  
If we imagine a 100MB mempool, then filling it with 30MB of transactions (sizewise = 100MB of usage) at a 100K sat/KB fee rate  will cost 30 BTC.   

In this case access to the network will be blocked for all txs less than 100k feerate for 5 hours while those transactions are mined anyway.  The additional gain the `rollingMinimumFeeRate` gives an attacker is another 7 hours until the decay has brought down the feerate to 50K.   

Since the attacker could have stopped anything under 50K feerate anyway for 10 hours by just issuing 60MB worth of transactions at that fee rate.  This attack is not significantly worse.

So I think 12 hours strikes about the right balance.