@jgarzik 2015-09-30
concept ACK - prefer 24-48 hours - will do some testing