@sdaftuar 2015-10-01
Reorgs should probably be handled differently -- I don't think it makes sense for eviction to take place when calling `AcceptToMemoryPool()` from `DisconnectBlock()`; instead perhaps we can just let the mempool grow during a reorg and trim it down to size at the end?