@TheBlueMatt 2015-10-01
Addressed a few nits...Things left to do:

- [ ] Figure out the decay constant/drop min fee to 0 when it gets near 0
- [ ] Steal code from #6557 to use CalculateDescendants to make the TrimToSize code simpler