@TheBlueMatt 2015-10-02
Removed more than a third of the lines in TrimToSize and removed some other code in mempool sorting :)