@TheBlueMatt 2015-10-02
Halflife set to:
if mempool is < max_mempool_size / 4:
    halflife = 3 hours
elif mempool < max_mempool_size / 2:
    halflife = 6 hours
else
    halflife = 12 hours.

When halflife is < minRelayTxFee (1000 satoshisPerKb), it is rounded down to 0.