@morcos 2015-10-02
+1 on the half life suggestions.