@morcos 2015-10-17
I've done some benchmarking.  I ran a historical simulation over 1 million transactions between July 6th and July 14th.  And I repeated this for 3 different code bases.  
- master as of 7/31 (before multindex and package tracking)
- master as of 10/16
- this pull

Results are below, I think you can see that the average is within measurement error.  The slight increase in the median is to be expected.  
All times in ms.

Txns accepted to the mempool |  pre-packages | master 10/16 | 6722
----------------------------------------- | ----------------- | ----------------- | --------
Average | 0.786 | 0.790 | 0.782
Median | 0.268 | 0.279 | 0.298
90th percentile | 1.04 | 0.91 | 0.93
99th percentile | 8.03 | 7.91 | 7.71
99.9th percentile | 45.8 | 45.7 | 43.1
max | 286 | 286 | 398

I broke down the timing a little bit more and it's clear that the vast majority of the outliers come from CheckInputs and HaveInputs.  For instance if you look at the 1k transactions between the 99.8th and 99.9th percentile. 

AcceptToMemoryPool | time in ms
----- | -----
Total | 40.9
CheckInputs | 32.5
HaveInputs | 6.8
remaining | 1.6

Although I did see 3 (out of 1M) transactions where the addUnchecked call took > 50ms in this pull.

These measurements were made with libsecp256k1 merged and default dbcache size.
6722 used a 100 MB mempool limit.